requirements.txt
request
pyrebase4
bme680
python-telegram-bot
spacy